Cleanup benchmarks
